Update services and package-version (#116)
* Rename services names

* temporary fix for two issues raised

* update gitignore

* update Services

* Update .gitignore

* remove models

* update package-lock version

* update
