+++ camunda-formio-tasklist-vue/src/services/index.ts
@@ -1,14 +1,14 @@
-export * from "./axios";
-export * from "./camunda-rest";
-export * from "./constants";
-export * from "./filterListFormatterService";
-export * from "./format-time";
-export * from "./formio-token";
-export * from "./formsflowai-api";
-export * from "./get-formio";
-export * from "./search-constants";
+export * from "./Auth";
+export * from "./Axios";
+export * from "./CamundaRest";
+export * from "./Constants";
+export * from "./FilterListFormatterService";
+export * from "./FormatTime";
+export * from "./FormioToken";
+export * from "./FormsFlowApi";
+export * from "./GetFormio";
+export * from "./SearchOptions";
 export * from "./SortOptions";
 export * from "./SocketIOServices";
 export * from "./LodashClone";
-export * from "./utils";
-export * from "./auth";
+export * from "./Utils";
